Remove examples from binary.  Size reduced from 19MiB to 10MiB.
The files are now loaded from the local resources/ directory.  This only
works for developers so these need to be packaged for end users somehow.

One possibility, assuming you want to keep the "Open Example" menu, is for
the program to look in the install directory for example files and populate
the menu from there.